[cmd & tests] cleanup unwrap.
