+++ block-relayer/src/metrics.rs
@@ -2,7 +2,7 @@ use once_cell::sync::Lazy;
 use starcoin_metrics::{register_int_gauge, IntGauge, Opts, PrometheusError};
 
 pub static BLOCK_RELAYER_METRICS: Lazy<BlockRelayerMetrics> =
-    Lazy::new(|| BlockRelayerMetrics::register().unwrap());
+    Lazy::new(|| BlockRelayerMetrics::register().expect("BlockRelayerMetrics register should ok."));
 
 #[derive(Clone)]
 pub struct BlockRelayerMetrics {
