+++ commons/decrypt/src/lib.rs
@@ -139,17 +139,4 @@ pub fn decrypt(secret: &[u8], encrypted: &[u8]) -> Result<Vec<u8>> {
 }
 
 #[cfg(test)]
-mod tests {
-    use crate::{decrypt, encrypt};
-
-    #[test]
-    fn test_encryption() {
-        let secret = "hello";
-        let plain = "world";
-        let encrypted = encrypt(secret.as_bytes(), plain.as_bytes());
-        assert_ne!(encrypted.as_slice(), plain.as_bytes());
-
-        let decrypted = decrypt(secret.as_bytes(), encrypted.as_slice()).unwrap();
-        assert_eq!(decrypted.as_slice(), plain.as_bytes());
-    }
-}
+mod tests;
