+++ network/api/src/peer_provider.rs
@@ -360,46 +360,3 @@ impl PeerSelector {
             .collect()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::peer_provider::{PeerSelector, PeerStrategy};
-    use starcoin_crypto::HashValue;
-    use starcoin_types::peer_info::{PeerId, PeerInfo};
-    use starcoin_types::startup_info::{ChainInfo, ChainStatus};
-    use starcoin_types::U256;
-
-    fn mock_chain_status(total_difficulty: U256) -> ChainStatus {
-        let mut status = ChainStatus::random();
-        status.info.total_difficulty = total_difficulty;
-        status
-    }
-    #[test]
-    fn test_peer_selector() {
-        let peers = vec![
-            PeerInfo::new(
-                PeerId::random(),
-                ChainInfo::new(1.into(), HashValue::zero(), mock_chain_status(100.into())),
-            ),
-            PeerInfo::new(
-                PeerId::random(),
-                ChainInfo::new(1.into(), HashValue::zero(), mock_chain_status(99.into())),
-            ),
-            PeerInfo::new(
-                PeerId::random(),
-                ChainInfo::new(1.into(), HashValue::zero(), mock_chain_status(100.into())),
-            ),
-            PeerInfo::new(
-                PeerId::random(),
-                ChainInfo::new(1.into(), HashValue::zero(), mock_chain_status(1.into())),
-            ),
-        ];
-
-        let peer_selector = PeerSelector::new(peers, PeerStrategy::default());
-        let beat_selector = peer_selector.bests().unwrap();
-        assert_eq!(2, beat_selector.len());
-
-        let top_selector = peer_selector.top(3);
-        assert_eq!(3, top_selector.len());
-    }
-}
