+++ vm/types/src/unit_tests/mod.rs
@@ -1,5 +1,6 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+mod access_path_test;
 mod block_metadata_test;
 mod transaction_test;
